# packages/generated/api-client-csharp/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plerion
  version: 0.1.0
servers:
- url: /
security:
- oauth2:
  - openid
- bearerAuth: []
paths:
...
  /capture_sessions/{id}/reconstructions:
    get:
      operationId: get_capture_session_reconstructions
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  format: uuid
                  type: string
                  x-schema-name: get_capture_session_reconstructions_resp_200_application_json_item
                  nullable: false
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Capture Session Reconstructions
      tags:
      - Default
...
  /reconstructions:
    get:
      operationId: get_reconstructions
      parameters:
      - description: Optional list of Ids to filter by
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: uuid
            type: string
            x-schema-name: get_reconstructions_param_ids_item
            nullable: false
          type: array
          nullable: true
        style: form
      - description: Optional capture session Id to filter by
        explode: true
        in: query
        name: capture_session_id
        required: false
        schema:
          format: uuid
          nullable: true
          type: string
        style: form
      - description: Optional capture session name to filter by
        explode: true
        in: query
        name: capture_session_name
        required: false
        schema:
          type: string
          nullable: true
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ReconstructionRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstructions
      tags:
      - Default
    post:
      operationId: create_reconstruction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconstructionCreateWithOptions"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconstructionRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Reconstruction
      tags:
      - Default
  /reconstructions/{id}:
    delete:
      operationId: delete_reconstruction
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Reconstruction
      tags:
      - Default
    get:
      operationId: get_reconstruction
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconstructionRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction
      tags:
      - Default
  /reconstructions/{id}/manifest:
    get:
      operationId: get_reconstruction_manifest
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconstructionManifest"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Manifest
      tags:
      - Default
  /reconstructions/{id}/localization_map:
    get:
      operationId: get_reconstruction_localization_map
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                format: uuid
                title: Response Get Reconstruction Localization Map
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Localization Map
      tags:
      - Default
  /reconstructions/{id}/status:
    get:
      operationId: get_reconstruction_status
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                enum:
                - queued
                - pending
                - downloading
                - extracting_features
                - matching_features
                - reconstructing
                - training_opq_matrix
                - training_product_quantizer
                - uploading
                - succeeded
                - failed
                title: Response Get Reconstruction Status
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Status
      tags:
      - Default
  /reconstructions/{id}/points:
    get:
      operationId: get_reconstruction_points
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PointCloudPoint"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Points
      tags:
      - Default
  /reconstructions/{id}/points.ply:
    get:
      operationId: get_reconstruction_points3D_ply
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Points3D Ply
      tags:
      - Default
  /reconstructions/{id}/image_poses:
    get:
      operationId: get_reconstruction_image_poses
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          title: Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Transform"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Reconstruction Image Poses
      tags:
      - Default
...
components:
  schemas:
...
    Color:
      example:
        r: 0
        b: 1
        g: 6
      properties:
        r:
          title: R
          type: integer
        g:
          title: G
          type: integer
        b:
          title: B
          type: integer
      required:
      - b
      - g
      - r
      title: Color
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    OrchestrationStatus:
      enum:
      - queued
      - pending
      - running
      - succeeded
      - cancelled
      - failed
      title: OrchestrationStatus
      type: string
...
    PointCloudPoint:
      example:
        color:
          r: 0
          b: 1
          g: 6
        position:
          x: 5.637376656633329
          "y": 2.3021358869347655
          z: 7.061401241503109
      properties:
        position:
          $ref: "#/components/schemas/Vector3"
        color:
          $ref: "#/components/schemas/Color"
      required:
      - color
      - position
      title: PointCloudPoint
    Quaternion:
      example:
        w: 5.962133916683182
        x: 0.8008281904610115
        "y": 6.027456183070403
        z: 1.4658129805029452
      properties:
        x:
          title: X
          type: number
        "y":
          title: "Y"
          type: number
        z:
          title: Z
          type: number
        w:
          title: W
          type: number
      required:
      - w
      - x
      - "y"
      - z
      title: Quaternion
    ReconstructionCreate:
      example:
        capture_session_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        orchestration_status: queued
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        capture_session_id:
          format: uuid
          title: Capture Session Id
          type: string
        id:
          format: uuid
          nullable: true
          type: string
        orchestration_status:
          $ref: "#/components/schemas/OrchestrationStatus"
      required:
      - capture_session_id
      title: ReconstructionCreate
    ReconstructionCreateWithOptions:
      example:
        options:
          random_seed: 0
          ransac_max_error: 5.962133916683182
          triangulation_merge_max_reprojection_error: 9.301444243932576
          ransac_min_inlier_ratio: 5.637376656633329
          compression_opq_number_bits_per_subvector: 4
          bundle_adjustment_refine_sensor_from_rig: true
          bundle_adjustment_refine_principal_point: true
          compression_opq_number_of_subvectors: 2
          use_prior_position: true
          max_keypoints_per_image: 1
          compression_opq_number_of_training_iterations: 7
          neighbors_count: 6
          triangulation_complete_max_reprojection_error: 7.061401241503109
          mapper_filter_max_reprojection_error: 3.616076749251911
          triangulation_minimum_angle: 2.3021358869347655
          bundle_adjustment_refine_additional_params: true
          bundle_adjustment_refine_focal_length: true
          single_threaded: true
          rig_verification: true
        create:
          capture_session_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          orchestration_status: queued
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        create:
          $ref: "#/components/schemas/ReconstructionCreate"
        options:
          $ref: "#/components/schemas/ReconstructionOptions"
      required:
      - create
      title: ReconstructionCreateWithOptions
    ReconstructionManifest:
      example:
        options:
          random_seed: 0
          ransac_max_error: 5.962133916683182
          triangulation_merge_max_reprojection_error: 9.301444243932576
          ransac_min_inlier_ratio: 5.637376656633329
          compression_opq_number_bits_per_subvector: 4
          bundle_adjustment_refine_sensor_from_rig: true
          bundle_adjustment_refine_principal_point: true
          compression_opq_number_of_subvectors: 2
          use_prior_position: true
          max_keypoints_per_image: 1
          compression_opq_number_of_training_iterations: 7
          neighbors_count: 6
          triangulation_complete_max_reprojection_error: 7.061401241503109
          mapper_filter_max_reprojection_error: 3.616076749251911
          triangulation_minimum_angle: 2.3021358869347655
          bundle_adjustment_refine_additional_params: true
          bundle_adjustment_refine_focal_length: true
          single_threaded: true
          rig_verification: true
        metrics:
          track_length_50th_percentile: 9.301444243932576
          total_images: 0
          reprojection_pixel_error_50th_percentile: 2.3021358869347655
          average_keypoints_per_image: 5.637376656633329
          reprojection_pixel_error_90th_percentile: 7.061401241503109
          num_3d_points: 5
          registration_rate: 1.4658129805029452
          registered_images: 6
          percent_tracks_with_length_greater_than_or_equal_to_3: 3.616076749251911
        capture_id: capture_id
        error: error
        status: queued
      properties:
        capture_id:
          title: Capture Id
          type: string
        status:
          enum:
          - queued
          - pending
          - downloading
          - extracting_features
          - matching_features
          - reconstructing
          - training_opq_matrix
          - training_product_quantizer
          - uploading
          - succeeded
          - failed
          title: Status
          type: string
        error:
          nullable: true
          type: string
        options:
          $ref: "#/components/schemas/ReconstructionOptions"
        metrics:
          $ref: "#/components/schemas/ReconstructionMetrics"
      required:
      - capture_id
      - metrics
      - options
      - status
      title: ReconstructionManifest
    ReconstructionMetrics:
      example:
        track_length_50th_percentile: 9.301444243932576
        total_images: 0
        reprojection_pixel_error_50th_percentile: 2.3021358869347655
        average_keypoints_per_image: 5.637376656633329
        reprojection_pixel_error_90th_percentile: 7.061401241503109
        num_3d_points: 5
        registration_rate: 1.4658129805029452
        registered_images: 6
        percent_tracks_with_length_greater_than_or_equal_to_3: 3.616076749251911
      properties:
        total_images:
          nullable: true
          type: integer
        registered_images:
          nullable: true
          type: integer
        registration_rate:
          nullable: true
          type: number
        num_3d_points:
          nullable: true
          type: integer
        average_keypoints_per_image:
          nullable: true
          type: number
        reprojection_pixel_error_50th_percentile:
          nullable: true
          type: number
        reprojection_pixel_error_90th_percentile:
          nullable: true
          type: number
        track_length_50th_percentile:
          nullable: true
          type: number
        percent_tracks_with_length_greater_than_or_equal_to_3:
          nullable: true
          type: number
      title: ReconstructionMetrics
    ReconstructionOptions:
      example:
        random_seed: 0
        ransac_max_error: 5.962133916683182
        triangulation_merge_max_reprojection_error: 9.301444243932576
        ransac_min_inlier_ratio: 5.637376656633329
        compression_opq_number_bits_per_subvector: 4
        bundle_adjustment_refine_sensor_from_rig: true
        bundle_adjustment_refine_principal_point: true
        compression_opq_number_of_subvectors: 2
        use_prior_position: true
        max_keypoints_per_image: 1
        compression_opq_number_of_training_iterations: 7
        neighbors_count: 6
        triangulation_complete_max_reprojection_error: 7.061401241503109
        mapper_filter_max_reprojection_error: 3.616076749251911
        triangulation_minimum_angle: 2.3021358869347655
        bundle_adjustment_refine_additional_params: true
        bundle_adjustment_refine_focal_length: true
        single_threaded: true
        rig_verification: true
      properties:
        random_seed:
          nullable: true
          type: integer
        single_threaded:
          nullable: true
          type: boolean
        neighbors_count:
          nullable: true
          type: integer
        max_keypoints_per_image:
          nullable: true
          type: integer
        ransac_max_error:
          nullable: true
          type: number
        ransac_min_inlier_ratio:
          nullable: true
          type: number
        use_prior_position:
          nullable: true
          type: boolean
        rig_verification:
          nullable: true
          type: boolean
        triangulation_minimum_angle:
          nullable: true
          type: number
        triangulation_complete_max_reprojection_error:
          nullable: true
          type: number
        triangulation_merge_max_reprojection_error:
          nullable: true
          type: number
        mapper_filter_max_reprojection_error:
          nullable: true
          type: number
        bundle_adjustment_refine_sensor_from_rig:
          nullable: true
          type: boolean
        bundle_adjustment_refine_focal_length:
          nullable: true
          type: boolean
        bundle_adjustment_refine_principal_point:
          nullable: true
          type: boolean
        bundle_adjustment_refine_additional_params:
          nullable: true
          type: boolean
        compression_opq_number_of_subvectors:
          nullable: true
          type: integer
        compression_opq_number_bits_per_subvector:
          nullable: true
          type: integer
        compression_opq_number_of_training_iterations:
          nullable: true
          type: integer
      title: ReconstructionOptions
    ReconstructionRead:
      example:
        capture_session_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        updated_at: 2000-01-23T04:56:07.000+00:00
        orchestration_status: queued
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        capture_session_id:
          format: uuid
          title: Capture Session Id
          type: string
        id:
          format: uuid
          title: Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        orchestration_status:
          $ref: "#/components/schemas/OrchestrationStatus"
      required:
      - capture_session_id
      - created_at
      - id
      - orchestration_status
      - updated_at
      title: ReconstructionRead
...
    Transform:
      example:
        rotation:
          w: 5.962133916683182
          x: 0.8008281904610115
          "y": 6.027456183070403
          z: 1.4658129805029452
        position:
          x: 5.637376656633329
          "y": 2.3021358869347655
          z: 7.061401241503109
      properties:
        position:
          $ref: "#/components/schemas/Vector3"
        rotation:
          $ref: "#/components/schemas/Quaternion"
      required:
      - position
      - rotation
      title: Transform
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Vector3:
      example:
        x: 5.637376656633329
        "y": 2.3021358869347655
        z: 7.061401241503109
      properties:
        x:
          title: X
          type: number
        "y":
          title: "Y"
          type: number
        z:
          title: Z
          type: number
      required:
      - x
      - "y"
      - z
      title: Vector3
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes: {}